Placed assert functions into a test util file
